:dizzy: Adding parameters and return values to coroutines
Now you can pass parameters (once, when initializing the coroutine) instead of having to use `user_data`, which is now removed. This is nice cause you don't have to access any of the `Coroutine` struct members.

The coroutine can now return values on yield and on their normal return statement. The caller code gets this as a result of run, as in `coroutine_value := run(*coroutine);`

Because the coroutines are now polymorphic, I removed the `current_coroutine` member of the context. I had a solution running with it, but it lost some type information so some checks and asserts had to happen at runtime, which was more error prone and I wasn't a fan. The downside of this is that yield now needs to take the pointer to the coroutine that's yielding. The coroutine procedure also needs to take the coroutine pointer itself (so it can pass it to yield). The overall solution is a bit more verbose because of this but also simpler because of the complexity of `context.current_coroutine` is gone.

With these changes I believe we should be able to even call coroutines from other coroutines, but I haven't tested it yet.