:comet: Parameters and Return Types Niceties
This avoids some of the verbosity when using void parameters or return types. Before, for a `Coroutine(void, void)` you would need to declare a procedure that was in the following form:

```
the_coroutine :: (coroutine : *Coroutine(void, void), parameter : void) -> void
{
    /* . . . */
    dummy : void;
    return dummy;
}
```

Now you can just do the following and a*void* all the voids and dummy parameters and return values

```
the_coroutine :: (coroutine : *Coroutine(void, void))
{
    /* . . . */
}
```

There's also a new overload of the init procedure that allows you to initialize a coroutine with void parameters without passing any parameters, in the same spirit as the change described above.

Finally, now coroutines default initialized to returning true if `is_done(*corroutine)` is called. If you haven't initialized the coroutine this feels like a more correct behaviour, in case you have one of this somewhere in memory and you want to check if it's done without worrying if it's initialized or not.