:carpentry_saw: Explicitly pushing the pointer to the local coroutine context.
Investigating the stack behaviour of the coroutines, at the time of writing some of the assert(...) code can get inlined and cause 4096 bytes of usage on the stack on it. Putting no_inline on assert_helper(...) removes this. Started a discussion to see if it's worth doing this in the standard library.