:carpentry_saw: Removing need for `#asm` and fixing POSIX
`#asm` might end up going away in its current form. Plus all we needed for it was just to glue towards our inlined x64 assembly, which always wanted to do stuff that the in-language `#asm` wouldn't allow, like messing with the stack, ret-ing, jmp-int, etc.

Sadly, to create pure x64 subroutines right now without wrapping them in a procedure is kind of annoying. So for the bit `platform_switch_context` we just copy the asm bytes to executable memory at runtime. Ideally we'd eventually be able to do #bytes at global scope and take the address of it directly into a procedure variable (or even constant!) More info about this in the @@TODOs in this change.